update thu mua dieu

+++ PurchasingManager/QuanLyThuMua/GlobalVariable.cs
@@ -11,8 +11,8 @@ namespace QuanLyThuMua
                  "; Min Pool Size=0;Max Pool Size=1000;Pooling=true; Connect Timeout=65535;");
 
         public static string PathFile { get; set; }
-        public static int SoDoMin { get; set; }
-        public static int SoDoMax { get; set; }
+        public static float SoDoMin { get; set; }
+        public static float SoDoMax { get; set; }
 
     }
 }
